requirements.txt
pytest
selene>=2.0.0b10
allure-pytest
